+++ reddwarf/utils/clustering.py
@@ -1,19 +1,42 @@
 import pandas as pd
 import numpy as np
-from sklearn.cluster import KMeans
+from reddwarf.sklearn.model_selection import GridSearchNonCV
+from reddwarf.sklearn.cluster import PolisKMeans
 from sklearn.metrics import silhouette_score
-from typing import List, Tuple, Optional
+from typing import List, Optional
 
+RangeLike = int | tuple[int, int] | list[int]
+
+def to_range(r: RangeLike) -> range:
+    """
+    Creates an inclusive range from a list, tuple, or int.
+
+    Examples:
+        ```
+        to_range(2) # [2, 3]
+        to_range([2, 5]) # [2, 3, 4, 5]
+        to_range((2, 5)) # [2, 3, 4, 5]
+        ```
+    """
+    if isinstance(r, int):
+        start = end = r
+    elif isinstance(r, (tuple, list)) and len(r) == 2:
+        start, end = r
+    else:
+        raise ValueError("Expected int or a 2-element tuple/list")
+
+    return range(start, end + 1)  # inclusive
 
 # TODO: Start passing init_centers based on /math/pca2 endpoint data,
 # and see how often we get the same clusters.
 def run_kmeans(
         dataframe: pd.DataFrame,
         n_clusters: int = 2,
+        init="k-means++",
         # TODO: Improve this type. 3d?
         init_centers: Optional[List] = None,
         random_state: Optional[int] = None,
-) -> Tuple[np.ndarray | None, np.ndarray]:
+) -> PolisKMeans:
     """
     Runs K-Means clustering on a 2D DataFrame of xy points, for a specific K,
     and returns labels for each row and cluster centers. Optionally accepts
@@ -22,86 +45,66 @@ def run_kmeans(
     Args:
         dataframe (pd.DataFrame): A dataframe with two columns (assumed `x` and `y`).
         n_clusters (int): How many clusters k to assume.
-        init_centers (List): A list of xy coordinates to use as initial center guesses.
+        init (string): The cluster initialization strategy. See `PolisKMeans` docs.
+        init_centers (List): A list of xy coordinates to use as initial center guesses. See `PolisKMeans` docs.
         random_state (int): Determines random number generation for centroid initialization. Use an int to make the randomness deterministic.
 
     Returns:
-        cluster_labels (np.ndarray | None): A list of zero-indexed labels for each row in the dataframe
-        cluster_centers (np.ndarray): A list of center coords for clusters.
+        kmeans (PolisKMeans): The estimator object returned from PolisKMeans.
     """
-    if init_centers:
-        # Pass an array of xy coords to see kmeans guesses.
-        init_arg = init_centers[:n_clusters]
-    else:
-        # Use the default strategy in sklearn.
-        init_arg = "k-means++"
     # TODO: Set random_state to a value eventually, so calculation is deterministic.
-    kmeans = KMeans(
+    kmeans = PolisKMeans(
         n_clusters=n_clusters,
         random_state=random_state,
-        init = init_arg, # type:ignore (because sklearn things it's just a string)
-        n_init="auto"
+        init=init,
+        init_centers=init_centers,
     ).fit(dataframe)
 
-    return kmeans.labels_, kmeans.cluster_centers_
+    return kmeans
 
 def find_optimal_k(
         projected_data: pd.DataFrame,
-        max_group_count: int = 5,
+        k_bounds: RangeLike = [2, 5],
+        init="k-means++",
         init_centers: Optional[List] = None,
         random_state: Optional[int] = None,
-        debug: bool = False,
-) -> Tuple[int, float, np.ndarray | None, np.ndarray | None]:
+) -> tuple[int, float, PolisKMeans | None]:
     """
     Use silhouette scores to find the best number of clusters k to assume to fit the data.
 
     Args:
         projected_data (pd.DataFrame): A dataframe with two columns (assumed `x` and `y`).
-        max_group_count (int): The max K number of groups to test for. (Default: 5)
+        k_bounds (RangeLike): An upper and low bound on n_clusters to test for. (Default: [2, 5])
         init_centers (List): A list of xy coordinates to use as initial center guesses.
         random_state (int): Determines random number generation for centroid initialization. Use an int to make the randomness deterministic.
-        debug (bool): Whether to print debug output. (Default: False)
 
     Returns:
         optimal_k (int): Ideal number of clusters.
         optimal_silhouette_score (float): Silhouette score for this K value.
-        optimal_cluster_labels (np.ndarray | None): A list of index labels assigned a group to each row in projected_date.
-        optimal_cluster_centers (np.ndarray | None): A list of xy centers for the optimal clusters.
+        optimal_kmeans (PolisKMeans | None): The optimal fitted estimator returned from PolisKMeans.
     """
-    K_RANGE = range(2, max_group_count+1)
-    k_best = 0 # Best K so far.
-    best_silhouette_score = -np.inf
-    best_cluster_labels = None
-    best_cluster_centers = None
+    param_grid = {
+        "n_clusters": to_range(k_bounds),
+    }
 
-    for k_test in K_RANGE:
-        cluster_labels, cluster_centers = run_kmeans(
-            dataframe=projected_data,
-            n_clusters=k_test,
-            init_centers=init_centers,
+    def scoring_function(estimator, X):
+        labels = estimator.fit_predict(X)
+        return silhouette_score(X, labels)
+
+    search = GridSearchNonCV(
+        param_grid=param_grid,
+        scoring=scoring_function,
+        estimator=PolisKMeans(
+            init=init, # strategy
+            init_centers=init_centers, # guesses
             random_state=random_state,
-        )
-        this_silhouette_score = silhouette_score(projected_data, cluster_labels)
-        if debug:
-            print(f"{k_test=}, {this_silhouette_score=}")
-        if this_silhouette_score >= best_silhouette_score:
-            k_best = k_test
-            best_silhouette_score = this_silhouette_score
-            best_cluster_labels = cluster_labels
-            best_cluster_centers = cluster_centers
+        ),
+    )
 
-    optimal_k = k_best
-    optimal_silhouette = best_silhouette_score
-    optimal_cluster_labels = best_cluster_labels
-    optimal_cluster_centers = best_cluster_centers
+    search.fit(projected_data)
 
-    return optimal_k, optimal_silhouette, optimal_cluster_labels, optimal_cluster_centers
+    best_k = search.best_params_['n_clusters']
+    best_silhouette_score = search.best_score_
+    best_kmeans = search.best_estimator_
 
-# TODO: Be smarter about this, and randomly generate extra centroid guesses via
-# best-practice. See Polis codebase for their way of generating.
-# Meaningless clusters are more likely in the center, so important to distribute.
-def pad_centroid_list_to_length(lst, size):
-    """
-    Pad a list of centroid guesses with [0,0] up to specific length.
-    """
-    return list(lst) + [[0., 0.]]*(size - len(lst))
+    return best_k, best_silhouette_score, best_kmeans